exersise/mnist/requirements.txt
absl-py==0.7.0
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
grpcio==1.19.0
h5py==2.8.0
Keras==2.2.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
mock==2.0.0
numpy==1.15.1
opencv-python==4.0.0.21
pbr==5.1.2
Pillow==5.4.1
protobuf==3.6.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
scipy==1.2.1
six==1.11.0
subprocess32==3.5.3
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-estimator==1.13.0
tensorflow-hub==0.1.1
tensorflowjs==0.8.0
termcolor==1.1.0
Werkzeug==0.14.1
